chore : back to prev struct
