+++ nextflix/pages/Movies/index.js
@@ -1,39 +1,56 @@
-import axios from "axios";
-import { useEffect,useState } from "react";
-import styled from "styled-components";
-import { getNowPlaying } from "../api/api";
+// import axios from 'axios';
+// import React, { useState,useEffect } from 'react';
+// import { use } from 'react';
 
-const MoviePoster = styled.img`
-    width: 40px;
-    height: 40px;
-`
+// // const Movies = () =>{
+// //     console.log(getMovies());
 
-export const getServerSideProps = async () =>{
-    const res =  await getNowPlaying();
-    const data = res.data;
+// //     return(
+// //         <div>
+// //             {/* {Movie.data} */}
+// //         </div>
+// //     );
+// // }
 
-  return { props: { data } }
-}
+// // export const getMovies = async () =>{
+// //     // const response = await fetch(`https://api.themoviedb.org/3/movie/now_playing?api_key=f85ba1745021cb0f98ac340407ad592b`);
+// //     const response = await axios.get(`https://api.themoviedb.org/3/movie/now_playing?api_key=f85ba1745021cb0f98ac340407ad592b`);
+// //     // const data = await response.json();
+// //     console.log(response);
+// //     // return data;
+// // }
 
 
-const Movies =({data}) =>{
-    const [movies,setMovies] = useState([]);
+// // const api = axios.create({
+// //     baseURL: 'https://api.themoviedb.org/3/movie/',
+// //     params: {
+// //       api_key: '761b201bf66269aa06a33c1d922810c2',
+// //       language: 'ko'
+// //     }
+// //   });
+  
+// //   export const movieApi = {
+// //     nowPlaying: () => api.get('now_playing'),
+// //     // popular: () => api.get('popular'),
+// //     // topRated: () => api.get('top_rated'),
+// //     // upcoming: () => api.get('upcoming'),
+// //   }
 
-    useEffect(()=>{
-        setMovies(data.results);
-        console.log(movies);
-    },[movies])
+// //   export const getMovies = async () =>{
+// //     const response = await  api.get('now_playing');
+// //     // console.log(movieApi);
+// //     return response.data.results;
+// //   }
 
-    return(
-    <div>
-    {movies.map((movie)=>(    
-        <div>
-        <img src={`https://image.tmdb.org/t/p/w185/${movie.poster_path}`}/>
-        </div>
-    ))}
-    </div>
-    );
+// export const getServerSideProps = async () =>{
+//     const res = await axios.get(`https://api.themoviedb.org/3/movie/now_playing?api_key=f85ba1745021cb0f98ac340407ad592b`)
+//      const data = res.data
 
-}
+//   return { props: { data } }
+// }
 
-export default Movies;
+// const Movies = ({movies}) =>{
+//     console.log(movies);
+// }
+
+// export default Movies;
